+++ main.py
@@ -1,4 +1,4 @@
-from core import run
+from corelight import run
 
 if __name__ == "__main__":
     run()
